refactor/requirements.txt
tinydb
pyyaml
pymongo
openapi3
python-slugify
